Update namespace to be Identity::Logging
