+++ lib/identity-logging.rb
@@ -1,5 +1,1 @@
-require 'identity-logging/version'
-require 'identity-logging/log_formatter'
-
-module IdentityLogging
-end
+require 'identity/logging'
