--- lib/identity-logging/version.rb
@@ -1,3 +1,0 @@
-module IdentityLogging
-  VERSION = '0.1.0'
-end
