+++ lib/identity/logging.rb
@@ -1,0 +1,7 @@
+require 'identity/logging/version'
+require 'identity/logging/log_formatter'
+
+module Identity
+  module Logging
+  end
+end
