+++ lib/identity/logging/version.rb
@@ -1,0 +1,5 @@
+module Identity
+  module Logging
+    VERSION = '0.1.0'
+  end
+end
