+++ spec/identity-logging_spec.rb
@@ -1,5 +1,5 @@
-RSpec.describe IdentityLogging do
+RSpec.describe Identity::Logging do
   it "has a version number" do
-    expect(IdentityLogging::VERSION).not_to be nil
+    expect(Identity::Logging::VERSION).not_to be nil
   end
 end
